rename tabswipe context. move file into context folder
