--- packages/app/src/contexts/TabSwipe.tsx
@@ -1,3 +1,0 @@
-import { createContext } from 'react';
-
-export const TabSwipeContext = createContext();
