+++ packages/app/src/contexts/TabSwipeContext/TabSwipeContext.tsx
@@ -1,0 +1,11 @@
+import { createContext, Dispatch, SetStateAction } from 'react';
+
+interface TabSwipeContextType {
+    canSwipe: boolean;
+    setCanSwipe: Dispatch<SetStateAction<boolean>>;
+}
+
+export const TabSwipeContext = createContext<TabSwipeContextType>({
+    canSwipe: true,
+    setCanSwipe: () => {}
+});
