+++ packages/app/src/contexts/index.ts
@@ -1,3 +1,4 @@
 import { Web3Context, Web3ContextProvider } from './Web3Context/Web3Context';
+import { TabSwipeContext } from './TabSwipeContext/TabSwipeContext';
 
-export { Web3Context, Web3ContextProvider };
+export { Web3Context, Web3ContextProvider, TabSwipeContext };
